биг беби теп лут

+++ Assets/_Project/Develop/Runtime/Configs/Gameplay/Loot/SoulShardLootConfig.cs
@@ -5,6 +5,6 @@ namespace Assets._Project.Develop.Runtime.Configs.Gameplay.Loot
     [CreateAssetMenu(menuName = "Configs/Gameplay/Loot/New Soul Shard Loot Config", fileName = "New Soul Shard Loot Config")]
     public class SoulShardLootConfig : LootConfig
     {
-        [field: SerializeField] public float ExperienceAmount { get; private set; }
+        [field: SerializeField] public int ExperienceAmount { get; private set; }
     }
 }
